backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
websockets==11.0.3
langchain==0.1.0
langchain-openai==0.0.2
python-dotenv==1.0.0
langchain-community==0.0.13
duckduckgo-search==4.1.1
langchain-core==0.1.9
requests==2.31.0
markdown==3.5.2
beautifulsoup4==4.12.3
python-jose==3.3.0 